Initial implementation sketch of internal architecture
The actual archipelago_rs AP client gets to live on its own thread,
and communicates with the rest of the mod via channels.

Currently location scouting panics on a ContinuationNotStarted
WebSocketError

The egui UI stuff has been disabled for now to focus on thinking
about the rest of the mod.

Co-authored-by: roseasromeo <11944660+roseasromeo@users.noreply.github.com>